+++ dataservices/models.py
@@ -299,21 +299,22 @@ class PopulationData(models.Model):
         unique_together = ('country', 'gender', 'year')
 
 
-class CommodityExports(TimeStampedModel):
-    root_code = models.CharField(null=True, blank=True, max_length=10)
-    commodity_code = models.CharField(null=True, blank=True, max_length=10)
-    commodity = models.CharField(blank=False, null=False, max_length=250)
+class UKTradeInGoodsByCountry(models.Model):
     country = models.ForeignKey(
         'dataservices.Country', verbose_name=_('Countries'), on_delete=models.SET_NULL, null=True
     )
-    direction = models.CharField(blank=True, null=True, max_length=50)
-    month = models.IntegerField(blank=True, null=True)
     year = models.IntegerField(blank=True, null=True)
-    value = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
+    quarter = models.PositiveSmallIntegerField(null=True, blank=True)
+    commodity_code = models.CharField(null=True, blank=True, max_length=10)
+    commodity_name = models.CharField(blank=False, null=False, max_length=250)
+    parent_code = models.CharField(null=True, blank=True, max_length=10)
+    imports = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
+    exports = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
+
+    objects = managers.UKTtradeInGoodsDataManager()
 
     class Meta:
-        verbose_name = "Country by commodity"
-        verbose_name_plural = "Country by commodity"
+        verbose_name = 'UK trade in goods by country'
 
 
 class UKTradeInServiceByCountry(models.Model):
@@ -321,13 +322,16 @@ class UKTradeInServiceByCountry(models.Model):
         'dataservices.Country', verbose_name=_('Countries'), on_delete=models.SET_NULL, null=True
     )
     year = models.IntegerField(null=True, blank=True)
-    quarter = models.IntegerField(null=True, blank=True)
+    quarter = models.PositiveSmallIntegerField(null=True, blank=True)
     service_code = models.CharField(null=True, blank=True, max_length=200)
     service_name = models.CharField(null=True, blank=True, max_length=200)
     imports = models.PositiveIntegerField(null=True, blank=True)
     exports = models.PositiveIntegerField(null=True, blank=True)
 
     objects = managers.UKTtradeInServicesDataManager()
+
+    class Meta:
+        verbose_name = 'UK trade in services by country'
 
 
 class UKTotalTradeByCountry(models.Model):