+++ dataservices/serializers.py
@@ -174,16 +174,22 @@ class PopulationDataSerializer(serializers.ModelSerializer):
         }
 
 
-class CommodityExportsSerializer(serializers.ModelSerializer):
-    country_name = serializers.CharField(source='country.name')
+class UKTopFiveGoodsExportsSerializer(serializers.ModelSerializer):
+    label = serializers.SerializerMethodField()
+    value = serializers.SerializerMethodField()
+
+    def get_label(self, obj):
+        return obj['label']
+
+    def get_value(self, obj):
+        return millions_to_currency_unit(obj['value'])
 
     class Meta:
-        model = models.CommodityExports
-        exclude = ['id', 'modified', 'country', 'created']
+        model = models.UKTradeInGoodsByCountry
+        fields = ['label', 'value']
 
 
-class UKTradeInServicesByCountrySerializer(serializers.ModelSerializer):
-    # country_name = serializers.CharField(source='country.name')
+class UKTopFiveServicesExportSerializer(serializers.ModelSerializer):
     label = serializers.SerializerMethodField()
     value = serializers.SerializerMethodField()