+++ dataservices/tests/conftest.py
@@ -133,7 +133,7 @@ def total_trade_records(countries):
 @pytest.fixture()
 def trade_in_services_records(countries):
     for idx, iso2 in enumerate(['DE', 'FR', 'CN']):
-        services = [
+        records = [
             {'code': '1', 'name': 'first', 'exports': 6, 'imports': 1},
             {'code': '2', 'name': 'second', 'exports': 5, 'imports': 1},
             {'code': '3', 'name': 'third', 'exports': 4, 'imports': 1},
@@ -143,13 +143,39 @@ def trade_in_services_records(countries):
         ]
         for year in [2020, 2021]:
             for quarter in [1, 2, 3, 4]:
-                for service in services:
+                for record in records:
                     models.UKTradeInServiceByCountry.objects.create(
                         country=countries[iso2],
                         year=year,
                         quarter=quarter,
-                        service_code=service['code'],
-                        service_name=service['name'],
-                        imports=service['imports'],
-                        exports=service['exports'],
+                        service_code=record['code'],
+                        service_name=record['name'],
+                        imports=record['imports'],
+                        exports=record['exports'],
+                    )
+
+
+@pytest.fixture()
+def trade_in_goods_records(countries):
+    for idx, iso2 in enumerate(['DE', 'FR', 'CN']):
+        records = [
+            {'code': '1', 'name': 'first', 'parent': None, 'exports': 6, 'imports': 1},
+            {'code': '2', 'name': 'second', 'parent': None, 'exports': 5, 'imports': 1},
+            {'code': '3', 'name': 'third', 'parent': None, 'exports': 4, 'imports': 1},
+            {'code': '4', 'name': 'fourth', 'parent': None, 'exports': 3, 'imports': 1},
+            {'code': '5', 'name': 'fifth', 'parent': None, 'exports': 2, 'imports': 1},
+            {'code': '6', 'name': 'last', 'parent': None, 'exports': 1, 'imports': 1},
+        ]
+        for year in [2020, 2021]:
+            for quarter in [1, 2, 3, 4]:
+                for record in records:
+                    models.UKTradeInGoodsByCountry.objects.create(
+                        country=countries[iso2],
+                        year=year,
+                        quarter=quarter,
+                        commodity_code=record['code'],
+                        commodity_name=record['name'],
+                        parent_code=record['parent'],
+                        imports=record['imports'],
+                        exports=record['exports'],
                     )