+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - GLS-186 - Dataservices addition for market trends
 - GLS-186 - Dataservices addition for trade highlights
 - GLS-232 - Ingest ONS data for UK total trade from Data Workspace
+- GLS-233 - Ingest ONS data for UK trade in goods from Data Workspace
 - GLS-234 - Ingest ONS data for UK trade in services from Data Workspace
 
 ### Bugs fixed